出品者以外の人が削除できないように比較分の導入

+++ app/controllers/items_controller.rb
@@ -34,10 +34,12 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    if @item.destroy
-      redirect_to root_path
-    else
-      render :show
+    if @item.user.id == current_user.id
+      if @item.destroy
+        redirect_to root_path
+      else
+        render :show
+      end
     end
   end
